MOVE enum Operator from Calulator.class to new file
