+++ src/main/java/calculator/Calculator.java
@@ -3,38 +3,30 @@ package calculator;
 import resource.StringResources;
 import util.StringUtils;
 
-import java.util.function.IntBinaryOperator;
-
 public class Calculator {
 
     private final String equation;
-    private final String[] elems;
+    private final String[] elements;
 
     public Calculator(String equation) throws IllegalArgumentException {
 
         this.equation = equation;
         verifyEmptyEquation();
 
-        elems = equation.split(" ");
-        verifyElemsLength();
+        elements = equation.split(" ");
+        verifyElementsLength();
     }
 
-    public int calculate() throws IllegalArgumentException {
+    public int calculate() throws IllegalArgumentException, ArithmeticException {
 
-        int result = Integer.parseInt(elems[0]);
+        int result = Integer.parseInt(elements[0]);
 
-        for (int i = 1; i < elems.length; i = i + 2) {
-            result = calculate(result, elems[i], elems[i + 1]);
+        for (int i = 1; i < elements.length; i = i + 2) {
+            Operator operator = Operator.of(elements[i]);
+            result = operator.calculate(result, Integer.parseInt(elements[i + 1]));
         }
 
         return result;
-    }
-
-    private int calculate(int result, String op, String num)
-            throws IllegalArgumentException, ArithmeticException {
-
-        Operator operator = Operator.of(op);
-        return operator.calculate(result, Integer.parseInt(num));
     }
 
     private void verifyEmptyEquation() throws IllegalArgumentException {
@@ -44,51 +36,10 @@ public class Calculator {
         }
     }
 
-    private void verifyElemsLength() {
+    private void verifyElementsLength() {
 
-        if (elems.length % 2 == 0) {
+        if (elements.length % 2 == 0) {
             throw new IllegalArgumentException(StringResources.ERR_INCORRECT_EQUATION);
-        }
-    }
-
-    private enum Operator {
-
-        PLUS((a, b) -> a + b),
-        SUBTRACT((a, b) -> a - b),
-        MULTIPLY((a, b) -> a * b),
-        DIVISION((a, b) -> {
-            if (b == 0) {
-                throw new ArithmeticException(StringResources.ERR_DIVIDED_BY_ZERO);
-            }
-            return a / b;
-        });
-
-        private final IntBinaryOperator intBinaryOperator;
-
-        Operator(IntBinaryOperator intBinaryOperator) {
-            this.intBinaryOperator = intBinaryOperator;
-        }
-
-        public static Operator of(String param) throws IllegalArgumentException {
-
-            char op = param.charAt(0);
-
-            switch (op) {
-                case '+':
-                    return Operator.PLUS;
-                case '-':
-                    return Operator.SUBTRACT;
-                case '*':
-                    return Operator.MULTIPLY;
-                case '/':
-                    return Operator.DIVISION;
-            }
-
-            throw new IllegalArgumentException(StringResources.ERR_INCORRECT_OPERATOR);
-        }
-
-        public int calculate(int a, int b) throws ArithmeticException {
-            return intBinaryOperator.applyAsInt(a, b);
         }
     }
 }
